% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bssdm-package.R
\docType{package}
\name{bssdm-package}
\alias{bssdm}
\alias{bssdm-package}
\title{bssdm: Biosecurity Species Distribution Modelling}
\description{
The bssdm package provides tools for species distribution modelling
with a focus on biosecurity applications. It includes implementations
of range bagging and Climatch algorithms, along with similarity
analysis methods (MESS, ExDet) and model validation tools.
}
\section{Main functions}{

\itemize{
\item \code{\link[=rangebag]{rangebag()}}: Range bagging species distribution modelling
\item \code{\link[=climatch]{climatch()}}: Climatch species distribution modelling
\item \code{\link[=mess]{mess()}}: Multivariate Environmental Similarity Surfaces
\item \code{\link[=exdet]{exdet()}}: Extrapolation detection in environmental space
\item \code{\link[=boyce]{boyce()}}: Boyce index for model validation
}
}

\section{Model prediction}{

\itemize{
\item \code{\link[=predict.Rangebag]{predict.Rangebag()}}: Predict method for Rangebag models
\item \code{\link[=predict.Climatch]{predict.Climatch()}}: Predict method for Climatch models
}
}

\references{
Drake, J. M. (2015). Range bagging: a new method for ecological
niche modelling from presence-only data.
\emph{Journal of the Royal Society Interface}, 12(107), 20150086.
\doi{10.1098/rsif.2015.0086}

ABARES (2020). Climatch v2.0 User Manual. Canberra.
\url{https://climatch.cp1.agriculture.gov.au/}

Elith, J., Kearney, M., & Phillips, S. (2010). The art of modelling
range-shifting species. \emph{Methods in Ecology and Evolution}, 1(4), 330-342.
\doi{10.1111/j.2041-210X.2010.00036.x}

Mesgaran, M. B., Cousens, R. D., & Webber, B. L. (2014). Here be dragons:
a tool for quantifying novelty due to covariate range and correlation change
when projecting species distribution models. \emph{Diversity and Distributions},
20(10), 1147-1159. \doi{10.1111/ddi.12209}

Hirzel, A. H., Le Lay, G., Helfer, V., Randin, C., & Guisan, A. (2006).
Evaluating the ability of habitat suitability models to predict species
presences. \emph{Ecological Modelling}, 199(2), 142-152.
\doi{10.1016/j.ecolmodel.2006.05.017}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cebra-analytics/bssdm}
}

}
\author{
\strong{Maintainer}: Sean Haythorne \email{sean.haythorne@unimelb.edu.au}

Authors:
\itemize{
  \item James Camac \email{james.camac@unimelb.edu.au}
  \item John Baumgartner \email{john.baumgartner@unimelb.edu.au}
}

Other contributors:
\itemize{
  \item Blaise Petitpierre (Author of Boyce index implementation from ecospat package) [copyright holder]
  \item Frank Breiner (Author of Boyce index implementation from ecospat package) [copyright holder]
  \item Flavien Collart (Contributor to Boyce index implementation from ecospat package) [copyright holder]
}

}
\keyword{internal}
