% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Climatch-class.R
\docType{class}
\name{Climatch-class}
\alias{Climatch-class}
\title{S4 class to represent a climatch SDM.}
\description{
The model object (S4) class component for an implementation of the ABARES
Climatch species distribution modelling (SDM) method (ABARES, 2020).
}
\section{Slots}{

\describe{
\item{\code{method}}{SDM method: "climatch".}

\item{\code{algorithm}}{Algorithm: "euclidean" or "closest_standard_score".}

\item{\code{variables}}{List of climate (or environmental) variable names.}

\item{\code{sd}}{The standard deviation of each variable calculated via the climate
data (\emph{x}) or the \emph{sd_data} when provided.}

\item{\code{presence}}{The selected (nearest within range) climate data for each
occurrence point.}

\item{\code{coordinates}}{The coordinates for the selected climate data.}

\item{\code{as_score}}{Indication of whether to generate a score 0-10 or values 0-1.}
}}

\references{
ABARES (2020). Climatch v2.0 User Manual. Canberra.
\url{https://climatch.cp1.agriculture.gov.au/} Accessed: November 2021.
}
