% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hawkweed}
\alias{hawkweed}
\title{Cleaned global occurrences of \emph{Hieracium pilosella} (hawkweed)}
\format{
A data frame with columns:
\describe{
\item{lon}{Longitude (degrees).}
\item{lat}{Latitude (degrees).}
\item{verified}{Numeric flag from the cleaning workflow.}
}
}
\source{
GBIF.org (2026) GBIF Occurrence Download. \doi{10.15468/dl.q6q6fk}
}
\usage{
hawkweed
}
\description{
Species occurrence coordinates (WGS84) used in package examples. Records
were downloaded from GBIF and cleaned with \emph{CoordinateCleaner} (Zizka et al.,
2019).
}
\references{
Zizka, A., Silvestro, D., Andermann, T., Azevedo, J.,
Duarte Ritter, C., Edler, D., Farooq, H., Herdean, A., Ariza, M., Scharn,
R., Svanteson, S., Wengstrom, N., Zizka, V., & Antonelli, A. (2019).
'CoordinateCleaner: standardized cleaning of occurrence records from
biological collection databases.' \emph{Methods in Ecology and Evolution}, 7.
\doi{10.1111/2041-210X.13152}
}
\keyword{datasets}
