% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mess.R
\name{mess}
\alias{mess}
\alias{mess.Raster}
\alias{mess.SpatRaster}
\alias{mess.data.frame}
\alias{mess.matrix}
\title{Calculate Multivariate Environmental Similarity}
\usage{
mess(x, ref, full = FALSE, filename = "", ...)

\method{mess}{Raster}(x, ref, full = FALSE, filename = "", ...)

\method{mess}{SpatRaster}(x, ref, full = FALSE, filename = "", ...)

\method{mess}{data.frame}(x, ref, full = FALSE, ...)

\method{mess}{matrix}(x, ref, full = FALSE, ...)
}
\arguments{
\item{x}{Climate (or environmental) data as a \code{terra::SpatRaster},
\code{raster::Raster*}, \code{list}, \code{data.frame}, or \code{matrix} where each
layer/element/column represents focal values of an environmental
variable.}

\item{ref}{A \code{data.frame}, \code{matrix}, or \code{list} where each column/element
represents reference values for an environmental variable (corresponding
to those given in \code{x}).}

\item{full}{Logical to indicate whether similarity values should be returned
for all variables. If \code{FALSE} (the default), then only the minimum
similarity scores across variables will be returned.}

\item{filename}{Optional filename for writing spatial raster output (i.e.,
the \code{mess} raster only). Default is "".}

\item{...}{Additional parameters.}
}
\value{
If \code{x} is a \code{SpatRaster} or \code{Raster*} object, this function returns
a list containing:
\itemize{
\item \code{mess}: a \code{SpatRaster} layer giving the minimum similarity
value across all variables for each location (i.e. the MESS);
\item \code{mess_by_variable}: a \code{SpatRaster} giving the environmental similarities
for each variable in \code{x} (only included when \code{full=TRUE});
\item \code{mod}: a factor \code{SpatRaster} layer indicating which variable was most
dissimilar to its reference range (i.e. the MoD map, Elith et al. 2010);
and
\item \code{mos}: a factor \code{SpatRaster} layer indicating which variable was most
similar to its reference range.
}

If \code{x} is a \code{list}, \code{matrix} or \code{data.frame}, the function will return a
list as above, but with multilayer and single layer \code{SpatRaster} objects
replaced by matrix and vector objects, respectively.
}
\description{
Calculate Multivariate Environmental Similarity and most dissimilar/similar
variables with respect to a reference dataset, for a set of environmental
variables.
}
\details{
\code{mess} uses the MESS algorithm described in Appendix S3 of Elith
et al. 2010.
}
\examples{
\dontrun{
library(geodata)
library(terra)
library(tmap)
bio <- worldclim_global("bio", res = 10, path = tempdir())
aus <- gadm("AUS", level = 0, resolution = 2, path = tempdir())
occ <- spatSample(aus, size = 100, method = "random")
ref <- terra::extract(bio, occ, ID = FALSE)
m <- mess(bio, ref, full = TRUE)

# Plot outputs
tmap::tm_shape(m$mess) +
  tmap::tm_raster(
    col.scale = tmap::tm_scale_continuous(
      values = "matplotlib.rd_bu",
      midpoint = 0
    ),
    col.legend = tmap::tm_legend(title = "MESS")
  ) +
  tmap::tm_credits(
    "Positive values indicate similarity, negative values indicate dissimilarity", 
    position = tmap::tm_pos_out("center", "bottom")
  ) +
  tmap::tm_title("Multivariate Environmental Similarity")

tmap::tm_shape(m$mod) +
  tmap::tm_raster(
    col.scale = tmap::tm_scale_categorical(
      values = "20",
      n.max = nrow(levels(m$mod)[[1]])
    ),
    col.legend = tmap::tm_legend(title = "Most dissimilar\ncovariate (MoD)"),
    col.chart = tmap::tm_chart_donut()
  ) +
  tmap::tm_title("Most dissimilar covariate (MoD)")

tmap::tm_shape(m$mos) +
  tmap::tm_raster(
    col.scale = tmap::tm_scale_categorical(
      values = "20",
      n.max = nrow(levels(m$mos)[[1]])
    ),
    col.legend = tmap::tm_legend(title = "Most similar\ncovariate (MoS)"),
    col.chart = tmap::tm_chart_donut()
  ) +
  tmap::tm_title("Most similar covariate (MoS)")
}
}
\references{
Elith, J., Kearney, M., and Phillips, S. (2010). The art of
modelling range-shifting species. \emph{Methods in Ecology and Evolution},
1: 330-342. \doi{10.1111/j.2041-210X.2010.00036.x}
}
\keyword{environment}
\keyword{maxent,}
\keyword{mess,}
\keyword{similarity,}
