% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boyce.R
\name{plot.boyce}
\alias{plot.boyce}
\title{Plot method for boyce objects}
\usage{
\method{plot}{boyce}(x, ...)
}
\arguments{
\item{x}{A boyce object}

\item{...}{Additional graphical parameters passed to plot}
}
\description{
Plots the predicted-to-expected ratio along habitat suitability values.
Grey points show all bins, black points show bins used in the correlation
calculation (after removing successive duplicates if applicable).
}
