% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boyce.R
\name{print.boyce}
\alias{print.boyce}
\title{Print method for boyce objects}
\usage{
\method{print}{boyce}(x, ...)
}
\arguments{
\item{x}{A boyce object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for boyce objects
}
