% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rangebag-class.R
\docType{class}
\name{Rangebag-class}
\alias{Rangebag-class}
\title{S4 class to represent a range bagging SDM.}
\description{
The model object (S4) class component for an implementation of the range
bagging species distribution modelling (SDM) method (Drake, 2015).
}
\section{Slots}{

\describe{
\item{\code{method}}{SDM method: "rangebag".}

\item{\code{variables}}{List of climate (or environmental) variable names.}

\item{\code{presence}}{The selected climate data corresponding to occurrences
points.}

\item{\code{coordinates}}{The coordinates for the selected climate data.}

\item{\code{ch_models}}{A list of convex hull models (vertices).}
}}

\references{
Drake, J. M. (2015). Range bagging: a new method for ecological
niche modelling from presence-only data.
\emph{Journal of the Royal Society Interface}, 12(107), 20150086.
\doi{10.1098/rsif.2015.0086}
}
