% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Climatch-class.R
\docType{class}
\name{Climatch-class}
\alias{Climatch-class}
\title{S4 class to represent a Climatch SDM.}
\description{
Description of Climatch class ...
}
\section{Slots}{

\describe{
\item{\code{method}}{SDM method: "climatch".}

\item{\code{algorithm}}{Algorithm: "euclidean" or "closest_standard_score".}

\item{\code{variables}}{List of climate (or environmental) variable names.}

\item{\code{sd}}{The standard deviation of each variable calculated via the climate data (\emph{x}) or the \emph{sd_data} when provided.}

\item{\code{presence}}{The selected (nearest within range) climate data for each occurrence point.}

\item{\code{coordinates}}{The coordinates for the selected climate data.}

\item{\code{as_score}}{Indication of whether to generate a score 0-10 or values 0-1.}
}}

