% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangebag.predict.R
\name{predict.Rangebag}
\alias{predict.Rangebag}
\title{Range bagging SDM predict method}
\usage{
\method{predict}{Rangebag}(
  object,
  x,
  raw_output = NULL,
  filename = "",
  parallel_cores = NULL,
  debug = FALSE,
  ...
)
}
\arguments{
\item{object}{A "Rangebag" model S4 object containing slots:
\describe{
\item{\code{method}}{SDM method: "rangebag".}
\item{\code{variables}}{List of climate (or environmental) variable
names.}
\item{\code{presence}}{The selected climate data corresponding to
occurrences points.}
\item{\code{coordinates}}{The coordinates for the selected climate
data.}
\item{\code{ch_models}}{A list of convex hull models (vertices).}
}}

\item{x}{Climate (or environmental) data with corresponding model variables
as a \code{terra::SpatRaster}, \code{raster::Raster*}, \code{data.frame},
or \code{matrix}.}

\item{raw_output}{Logical to indicate whether to return raw predicted
values (TRUE) or as an object (as per \emph{x}: FALSE). Default is NULL,
returning either raw values or a spatial raster (as per \emph{x}).}

\item{filename}{Optional filename for writing spatial raster output (only).
Default is "".}

\item{parallel_cores}{Optional number of cores available for parallel
processing, thus enable parallel processing. Default is NULL (serial).}

\item{debug}{Output additional debug information (memory block info, etc.).
Default is FALSE.}

\item{...}{Additional parameters.}
}
\value{
Predicted values as a raw vector or a \code{terra::SpatRaster},
\code{raster::Raster*}, \code{data.frame}, or \code{matrix} (as per
\emph{x}).
}
\description{
The model prediction component of an implementation of the range bagging
species distribution modelling (SDM) method (Drake, 2015).
}
\references{
Drake, J. M. (2015). Range bagging: a new method for ecological
niche modelling from presence-only data.
\emph{Journal of the Royal Society Interface}, 12(107), 20150086.
\doi{10.1098/rsif.2015.0086}
}
