% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rangebag-class.R
\docType{class}
\name{Rangebag-class}
\alias{Rangebag-class}
\title{S4 class to represent a Range Bagging SDM.}
\description{
Description of Rangebag class ...
}
\section{Slots}{

\describe{
\item{\code{method}}{SDM method: "rangebag".}

\item{\code{variables}}{List of climate (or environmental) variable names.}

\item{\code{presence}}{The selected climate data corresponding to occurrences points.}

\item{\code{coordinates}}{The coordinates for the selected climate data.}

\item{\code{ch_models}}{A list of convex hull models (vertices).}
}}

